hackathons/agents/world-monitor-agent/pyproject.toml
[tool.poetry]
name = "world-monitor-agent"
version = "0.1.0"
description = "World intelligence agent - fetches news, social, finance data and sells via Nevermined"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
strands-agents = {version = ">=1.0.0", extras = ["anthropic", "a2a"]}
payments-py = {version = ">=1.3.3", extras = ["strands"]}
fastapi = "^0.120.0"
uvicorn = ">=0.34.2,<1.0.0"
httpx = "^0.28.0"
anthropic = ">=0.40.0"
python-dotenv = "^1.0.0"
sse-starlette = "^2.0.0"
apify-client = "^1.8.0"
exa-py = "^1.0.0"

[tool.poetry.scripts]
web = "src.web:main"
agent-a2a = "src.agent_a2a:main"
setup = "src.setup:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
